backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
langchain==0.1.4
langchain-community==0.0.16
dashscope==1.14.1
serpapi==0.1.5
requests==2.31.0
pydantic==2.5.3
